+++ backend/src/baserow/contrib/database/fields/models.py
@@ -118,6 +118,9 @@ class SelectOption(models.Model):
     class Meta:
         ordering = ('order', 'id',)
 
+    def __str__(self):
+        return self.value
+
 
 class TextField(Field):
     text_default = models.CharField(